Streaming data in chunks and serializer improvements.
- Distributed runtime was limited to sending messages of maximum
size INT_MAX for each message (both for MPI and gRPC). With this
commit we are no longer limited by this and we serialize data into
chunks which are then sent to the workers, allowing us to send bigger
messages.
- Added a new command line arg to set maximum chunk size for distr
messages (default ~2GB).
- Use pointer for holding buffer inside serialization iterator.
- DaphneSerializer std::vector buffer, used reserve() and
capacity() methods to manage available space. However, we should not write to
vector's memory (even if reserved) without first resizing the buffer. Therefore replaced
reserve() calls with resize(). Similar replacements for MPI and gRPC worker.
- Due to resize() use, it's best to use the smallest possible chunk size.
- Added test case for distributed chunked messages.